Multiple changes
fix merge pom's conflict on release finish using the option
useSnapshotInRelease

change useSnapshot to allow its use in command line in batch mode

fix issue with develop version when the hotfix version is superior -
Develop one must be set to the hotfix one +1.

added new supportVersion for batch mode to be used when we want to set
the support version to a new version

added new property supportBranchName for batch mode so we can customize
the support branch name instead of using the tag by default

New parameter: useSupportHasDevelop - if a support branch exists we can
have the situation where the support is being used as a concurrent
development branch, for a different client for example, in parallel with
the development branch. In this case the version is normally a SNAPSHOT
and each time we do releases we want to increment the SNAPSHOT version.